backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
python-dotenv==1.2.2
deepagents==0.4.11
langchain-mongodb==0.11.0
pymongo==4.16.0
langchain-openai==1.1.11